functionality to delete a podcast
removes all trace of all the episodes in the podcast. the implementation
should be revisited. it abuses the multi-select functionality in the
gridview, even though we only want one.

+++ app/src/main/java/net/x4a42/volksempfaenger/data/entity/skippedepisode/SkippedEpisodeDaoWrapper.java
@@ -23,8 +23,11 @@ public class SkippedEpisodeDaoWrapper
 
     public void delete(Episode episode)
     {
-        SkippedEpisode skippedEpisode = getForEpisode(episode).get(0);
-        dao.delete(skippedEpisode);
+        if (hasEpisode(episode))
+        {
+            SkippedEpisode skippedEpisode = getForEpisode(episode).get(0);
+            dao.delete(skippedEpisode);
+        }
     }
 
     public SkippedEpisode create(Episode episode)
